Create README.md
This project implements algorithms for single and multiple freight train scheduling in a non-linear railway system shared with passenger trains. It adopts a heuristic and greedy paradigm to yield the optimal result (shortest travel time of freight train). A GUI is also included, which displays the velocity profiles of freight trains and passenger trains, a simplified graph representation for the railroad system, as well as schedule results in table format.